feat: R-5 dynamic template menu from domain plugins
Extract the hardcoded Examples dropdown from AppLayout into a new
TemplateMenu component that dynamically builds a hierarchical menu
(domain category → templates) from domainStore registered plugins.

- New: src/components/layout/TemplateMenu.tsx
- Changed: AppLayout.tsx now uses <TemplateMenu /> instead of inline menu
- No new dependencies needed

Closes #6
